[#934] Update references to renamed files
* Updates all imports to reference renamed fixture files
* Updates colorUtils.ts imports to color-utils.ts
* Removes extra blank line in window.ts documentation

Updates file naming to follow consistent patterns across the codebase
and removes unused code to improve maintainability.
