+++ src/plugins/pinia/logoutPlugin.ts
@@ -47,66 +47,3 @@ export function logoutPlugin(context: PiniaPluginContext) {
     };
   }
 }
-
-/**
-// Alternate implementation using store state
-if (!context.store.$logout) {
-  context.store.$logout = function () {
-    // Get all stores and reset them
-    const stores = Array.from(context.pinia.state.value).map(([id]) =>
-      context.pinia._s.get(id)
-    ).filter(Boolean);
-
-    // Reset all stores
-    stores.forEach(store => {
-      if (typeof store.$reset === 'function') {
-        store.$reset();
-      }
-    });
-
-    // Clear cookies
-    const COOKIES_TO_CLEAR = ['sess', 'locale'];
-    COOKIES_TO_CLEAR.forEach(deleteCookie);
-
-    // Clear browser storage
-    sessionStorage.clear();
-
-    // Reset window state
-    if ('cust' in window) window.cust = null;
-    if ('authenticated' in window) window.authenticated = false;
-
-    // Special handling for auth store if it exists
-    const authStore = context.pinia._s.get('auth');
-    if (authStore && '$stopAuthCheck' in authStore) {
-      authStore.$stopAuthCheck();
-    }
-
-    // Clear all Pinia state
-    context.pinia.state.value = {};
-  };
-}
-
-// Or go event based
-export const LOGOUT_EVENTS = {
-  BEFORE_LOGOUT: 'before-logout',
-  AFTER_LOGOUT: 'after-logout',
-  LOGOUT_ERROR: 'logout-error'
-} as const;
-
-export function logoutPlugin({ emitter }: { emitter?: EventEmitter } = {}) {
-  return ({ store }: PiniaPluginContext) => {
-    if (!store.$logout) {
-      store.$logout = async function() {
-        try {
-          emitter?.emit(LOGOUT_EVENTS.BEFORE_LOGOUT);
-          // Logout logic
-          emitter?.emit(LOGOUT_EVENTS.AFTER_LOGOUT);
-        } catch (error) {
-          emitter?.emit(LOGOUT_EVENTS.LOGOUT_ERROR, error);
-          throw error;
-        }
-      };
-    }
-  };
-}
-*/
