+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 // utils/index.ts
-export * from './colorUtils';
+export * from './color-utils';
 export * from './format';
 export * from './parse';
 export * from './redirect';
