+++ tests/unit/vue/composables/useDomainsManager.spec.ts
@@ -1,7 +1,7 @@
 // tests/unit/vue/composables/useDomainsManager.spec.ts
 import { useDomainsManager } from '@/composables/useDomainsManager';
 import { ApplicationError } from '@/schemas/errors';
-import { mockDomains, newDomainData } from '@tests/unit/vue/fixtures/domains';
+import { mockDomains, newDomainData } from '@tests/unit/vue/fixtures/domains.fixture';
 import { createPinia, setActivePinia } from 'pinia';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { ref } from 'vue';
